+++ codex/search/backend.py
@@ -1,5 +1,7 @@
 """Custom Haystack Search Backend."""
+from logging import getLogger
 from multiprocessing import cpu_count
+from queue import SimpleQueue
 from sys import maxsize
 from time import time
 
@@ -14,10 +16,11 @@ from whoosh.fields import NUMERIC
 from whoosh.qparser import FieldAliasPlugin, GtLtPlugin, OperatorsPlugin
 from whoosh.qparser.dateparse import DateParserPlugin
 from whoosh.support.charset import accent_map
-from whoosh.writing import BufferedWriter
+from whoosh.writing import MERGE_SMALL, BufferedWriter
 
 from codex.librarian.search.status import SearchIndexStatusTypes
 from codex.logger.logging import get_logger
+from codex.status_controller import StatusController
 from codex.worker_base import WorkerBaseMixin
 
 
@@ -100,20 +103,31 @@ class CodexSearchBackend(WhooshSearchBackend, WorkerBaseMixin):
         Not=r"(?i)(^|(?<=(\s|[()])))NOT(?=\s)",
         Require=r"(?i)(^|(?<=\s))REQUIRE(?=\s)",
     )
-    MULTISEGMENT_CUTOFF = 500
-    UPDATE_FINISH_TYPES = frozenset(
+    STATUS_FINISH_TYPES = frozenset(
         (
             SearchIndexStatusTypes.SEARCH_INDEX_CLEAR,
-            SearchIndexStatusTypes.SEARCH_INDEX_PREPARE,
-            SearchIndexStatusTypes.SEARCH_INDEX_COMMIT,
+            SearchIndexStatusTypes.SEARCH_INDEX_UPDATE,
         )
     )
+    WRITERARGS = {
+        "limitmb": maxsize,
+        "procs": cpu_count(),
+        # Bug in Whoosh means procs > 1 needs multisegment
+        # https://github.com/mchaput/whoosh/issues/35
+        "multisegement": True,
+    }
+    COMMITARGS = {"merge": True, "mergetype": MERGE_SMALL}
 
     def __init__(self, connection_alias, **connection_options):
         """Init worker queues."""
-        log_queue = connection_options.get("LOG_QUEUE")
-        librarian_queue = connection_options.get("LIBRARIAN_QUEUE")
-        self.init_worker(log_queue, librarian_queue)
+        log_queue = connection_options.pop("log_queue", None)
+        librarian_queue = connection_options.pop("librarian_queue", None)
+        if log_queue and librarian_queue:
+            self.init_worker(log_queue, librarian_queue)
+        else:
+            self.log = getLogger(self.__class__.__name__)
+            self.log.propagate = False
+            self.status_controller = StatusController(SimpleQueue(), SimpleQueue())
         super().__init__(connection_alias, **connection_options)
 
     def build_schema(self, fields):
@@ -156,29 +170,16 @@ class CodexSearchBackend(WhooshSearchBackend, WorkerBaseMixin):
         )
         self.parser.replace_plugin(self.OPERATORS_PLUGIN)
 
-    @classmethod
-    def get_writerargs(cls, num_objs):
-        """Use multiproc, multisegment writing only for large loads."""
-        writerargs = {
-            "limitmb": maxsize,
-        }
-        if num_objs > cls.MULTISEGMENT_CUTOFF:
-            # Bug in Whoosh means procs > 1 needs multisegment
-            # https://github.com/mchaput/whoosh/issues/35
-            writerargs["procs"] = cpu_count()
-            writerargs["multisegment"] = cpu_count() > 1
-        return writerargs
-
     def update(self, index, iterable, commit=True):
         """Update index, but with writer options."""
         start = since = time()
         try:
             num_objs = len(iterable)
             statuses = {
-                SearchIndexStatusTypes.SEARCH_INDEX_PREPARE: {
+                SearchIndexStatusTypes.SEARCH_INDEX_UPDATE: {
                     "total": num_objs,
                 },
-                SearchIndexStatusTypes.SEARCH_INDEX_COMMIT: {"name": f"({num_objs})"},
+                # SearchIndexStatusTypes.SEARCH_INDEX_COMMIT: {},
             }
             self.status_controller.start_many(statuses)
             if not self.setup_complete:
@@ -186,9 +187,14 @@ class CodexSearchBackend(WhooshSearchBackend, WorkerBaseMixin):
 
             self.index = self.index.refresh()
 
-            writerargs = self.get_writerargs(num_objs)
-            # writer = AsyncWriter(self.index, writerargs=writerargs)
-            writer = BufferedWriter(self.index, limit=1000, writerargs=writerargs)
+            # writer = AsyncWriter(self.index, writerargs=self.WRITERARGS)
+            writer = BufferedWriter(
+                self.index,
+                limit=maxsize,
+                period=60 * 5,
+                writerargs=self.WRITERARGS,
+                commitargs=self.COMMITARGS,
+            )
 
             for obj_count, obj in enumerate(iterable):
                 try:
@@ -221,43 +227,33 @@ class CodexSearchBackend(WhooshSearchBackend, WorkerBaseMixin):
                             },
                         )
                 since = self.status_controller.update(
-                    SearchIndexStatusTypes.SEARCH_INDEX_PREPARE,
+                    SearchIndexStatusTypes.SEARCH_INDEX_UPDATE,
                     obj_count,
                     num_objs,
                     since=since,
                 )
-
-            elapsed_time = time() - start
-            until = start + 1
-            self.status_controller.finish(
-                SearchIndexStatusTypes.SEARCH_INDEX_PREPARE, until=until
-            )
-            elapsed = naturaldelta(elapsed_time)
-            cps = int(num_objs / elapsed_time)
-            self.log.info(
-                f"Search engine prepared {num_objs} comics for commit"
-                f" in {elapsed} at {cps} comics per second."
-            )
-
             prepare_start = time()
-            # if len(iterable) > 0:
-            #    # For now, commit no matter what, as we run into locking issues
-            #    # otherwise.
-            #    if commit:
-            #        writer.commit(merge=True)
-            #    if writer.ident is not None:
-            #        writer.join()
+            if len(iterable) > 1:
+                writer.commit()
+            else:
+                writer.cancel()
             writer.close()
+
             elapsed_time = time() - prepare_start
             elapsed = naturaldelta(elapsed_time)
             cps = int(num_objs / elapsed_time)
             self.log.info(
-                f"Search engine committed index segment of {num_objs} comics"
+                f"Search engine updated {num_objs} comics"
                 f" in {elapsed} at {cps} comics per second."
             )
+            until = prepare_start + 1
+            self.status_controller.finish(
+                SearchIndexStatusTypes.SEARCH_INDEX_UPDATE, until=until
+            )
+
         finally:
             until = start + 1
-            self.status_controller.finish_many(self.UPDATE_FINISH_TYPES, until=until)
+            self.status_controller.finish_many(self.STATUS_FINISH_TYPES, until=until)
 
     def clear(self, models=None, commit=True):
         """Clear index with codex status messages."""
@@ -269,4 +265,53 @@ class CodexSearchBackend(WhooshSearchBackend, WorkerBaseMixin):
             until = start + 1
             self.status_controller.finish(
                 SearchIndexStatusTypes.SEARCH_INDEX_CLEAR, until=until
+            )
+
+    def remove_batch(self, doc_ids):
+        """Remove a large batch of doc ids from the index."""
+        num_doc_ids = len(doc_ids)
+        self.status_controller.start(
+            SearchIndexStatusTypes.SEARCH_INDEX_REMOVE, num_doc_ids
+        )
+        start = since = time()
+        try:
+            if not self.setup_complete:
+                self.setup()
+
+            self.index = self.index.refresh()
+            writer = BufferedWriter(
+                self.index,
+                limit=maxsize,
+                period=60 * 5,
+                writerargs=self.WRITERARGS,
+                commitargs=self.COMMITARGS,
+            )
+            count = 0
+            for count, doc_id in enumerate(doc_ids):
+                writer.delete_document(doc_id)
+                since = self.status_controller.update(
+                    SearchIndexStatusTypes.SEARCH_INDEX_REMOVE,
+                    count,
+                    num_doc_ids,
+                    since=since,
+                )
+
+            if len(doc_ids) > 1:
+                writer.commit()
+            else:
+                writer.cancel()
+            writer.close()
+
+            elapsed_time = time() - start
+
+            elapsed = naturaldelta(elapsed_time)
+            cps = int(count / elapsed_time)
+            self.log.info(
+                f"Search engine removed {count} ghosts from the index"
+                f" in {elapsed} at {cps} per second."
+            )
+        finally:
+            until = start + 1
+            self.status_controller.finish(
+                SearchIndexStatusTypes.SEARCH_INDEX_REMOVE, until=until
             )